+++ optics/psf2d.m
@@ -18,3 +18,5 @@ function I = psf2d(microscopeType, x, y, varargin)
     [xx, yy] = meshgrid(x, y);
     r = sqrt(xx.^2 + yy.^2);
     I = lateralIntensityDist(r, microscopeType, varargin{:});
+
+    I = I / max(I(:));